+++ sdk/dataprotection/arm-dataprotection/src/dataProtectionClient.ts
@@ -26,6 +26,10 @@ import {
   BackupPoliciesImpl,
   BackupInstancesImpl,
   RecoveryPointsImpl,
+  SecondaryRPsImpl,
+  CrossRegionRestoreJobImpl,
+  CrossRegionRestoreJobsImpl,
+  BackupInstancesExtensionRoutingImpl,
   JobsImpl,
   RestorableTimeRangesImpl,
   ExportJobsImpl,
@@ -46,6 +50,10 @@ import {
   BackupPolicies,
   BackupInstances,
   RecoveryPoints,
+  SecondaryRPs,
+  CrossRegionRestoreJob,
+  CrossRegionRestoreJobs,
+  BackupInstancesExtensionRouting,
   Jobs,
   RestorableTimeRanges,
   ExportJobs,
@@ -59,7 +67,7 @@ import { DataProtectionClientOptionalParams } from "./models";
 export class DataProtectionClient extends coreClient.ServiceClient {
   $host: string;
   apiVersion: string;
-  subscriptionId: string;
+  subscriptionId?: string;
 
   /**
    * Initializes a new instance of the DataProtectionClient class.
@@ -71,12 +79,26 @@ export class DataProtectionClient extends coreClient.ServiceClient {
     credentials: coreAuth.TokenCredential,
     subscriptionId: string,
     options?: DataProtectionClientOptionalParams
+  );
+  constructor(
+    credentials: coreAuth.TokenCredential,
+    options?: DataProtectionClientOptionalParams
+  );
+  constructor(
+    credentials: coreAuth.TokenCredential,
+    subscriptionIdOrOptions?: DataProtectionClientOptionalParams | string,
+    options?: DataProtectionClientOptionalParams
   ) {
     if (credentials === undefined) {
       throw new Error("'credentials' cannot be null");
     }
-    if (subscriptionId === undefined) {
-      throw new Error("'subscriptionId' cannot be null");
+
+    let subscriptionId: string | undefined;
+
+    if (typeof subscriptionIdOrOptions === "string") {
+      subscriptionId = subscriptionIdOrOptions;
+    } else if (typeof subscriptionIdOrOptions === "object") {
+      options = subscriptionIdOrOptions;
     }
 
     // Initializing default values for options
@@ -88,7 +110,7 @@ export class DataProtectionClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-dataprotection/1.1.0`;
+    const packageDetails = `azsdk-js-arm-dataprotection/2.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -141,7 +163,7 @@ export class DataProtectionClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-01-01";
+    this.apiVersion = options.apiVersion || "2023-06-01-preview";
     this.backupVaults = new BackupVaultsImpl(this);
     this.operationResult = new OperationResultImpl(this);
     this.operationStatus = new OperationStatusImpl(this);
@@ -159,6 +181,12 @@ export class DataProtectionClient extends coreClient.ServiceClient {
     this.backupPolicies = new BackupPoliciesImpl(this);
     this.backupInstances = new BackupInstancesImpl(this);
     this.recoveryPoints = new RecoveryPointsImpl(this);
+    this.secondaryRPs = new SecondaryRPsImpl(this);
+    this.crossRegionRestoreJob = new CrossRegionRestoreJobImpl(this);
+    this.crossRegionRestoreJobs = new CrossRegionRestoreJobsImpl(this);
+    this.backupInstancesExtensionRouting = new BackupInstancesExtensionRoutingImpl(
+      this
+    );
     this.jobs = new JobsImpl(this);
     this.restorableTimeRanges = new RestorableTimeRangesImpl(this);
     this.exportJobs = new ExportJobsImpl(this);
@@ -208,6 +236,10 @@ export class DataProtectionClient extends coreClient.ServiceClient {
   backupPolicies: BackupPolicies;
   backupInstances: BackupInstances;
   recoveryPoints: RecoveryPoints;
+  secondaryRPs: SecondaryRPs;
+  crossRegionRestoreJob: CrossRegionRestoreJob;
+  crossRegionRestoreJobs: CrossRegionRestoreJobs;
+  backupInstancesExtensionRouting: BackupInstancesExtensionRouting;
   jobs: Jobs;
   restorableTimeRanges: RestorableTimeRanges;
   exportJobs: ExportJobs;