+++ sdk/dataprotection/arm-dataprotection/src/models/parameters.ts
@@ -20,10 +20,15 @@ import {
   BackupInstanceResource as BackupInstanceResourceMapper,
   TriggerBackupRequest as TriggerBackupRequestMapper,
   ValidateForBackupRequest as ValidateForBackupRequestMapper,
+  CrossRegionRestoreRequestObject as CrossRegionRestoreRequestObjectMapper,
+  ValidateCrossRegionRestoreRequestObject as ValidateCrossRegionRestoreRequestObjectMapper,
   AzureBackupRehydrationRequest as AzureBackupRehydrationRequestMapper,
   AzureBackupRestoreRequest as AzureBackupRestoreRequestMapper,
   SyncBackupInstanceRequest as SyncBackupInstanceRequestMapper,
   ValidateRestoreRequestObject as ValidateRestoreRequestObjectMapper,
+  FetchSecondaryRPsRequestParameters as FetchSecondaryRPsRequestParametersMapper,
+  CrossRegionRestoreJobRequest as CrossRegionRestoreJobRequestMapper,
+  CrossRegionRestoreJobsRequest as CrossRegionRestoreJobsRequestMapper,
   AzureBackupFindRestorableTimeRangesRequest as AzureBackupFindRestorableTimeRangesRequestMapper,
   ResourceGuardResource as ResourceGuardResourceMapper,
   PatchResourceGuardInput as PatchResourceGuardInputMapper,
@@ -58,7 +63,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-01-01",
+    defaultValue: "2023-06-01-preview",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -214,20 +219,30 @@ export const parameters7: OperationParameter = {
 
 export const parameters8: OperationParameter = {
   parameterPath: "parameters",
-  mapper: AzureBackupRehydrationRequestMapper
+  mapper: CrossRegionRestoreRequestObjectMapper
 };
 
 export const parameters9: OperationParameter = {
   parameterPath: "parameters",
-  mapper: AzureBackupRestoreRequestMapper
+  mapper: ValidateCrossRegionRestoreRequestObjectMapper
 };
 
 export const parameters10: OperationParameter = {
   parameterPath: "parameters",
-  mapper: SyncBackupInstanceRequestMapper
+  mapper: AzureBackupRehydrationRequestMapper
 };
 
 export const parameters11: OperationParameter = {
+  parameterPath: "parameters",
+  mapper: AzureBackupRestoreRequestMapper
+};
+
+export const parameters12: OperationParameter = {
+  parameterPath: "parameters",
+  mapper: SyncBackupInstanceRequestMapper
+};
+
+export const parameters13: OperationParameter = {
   parameterPath: "parameters",
   mapper: ValidateRestoreRequestObjectMapper
 };
@@ -263,6 +278,33 @@ export const recoveryPointId: OperationURLParameter = {
   }
 };
 
+export const parameters14: OperationParameter = {
+  parameterPath: "parameters",
+  mapper: FetchSecondaryRPsRequestParametersMapper
+};
+
+export const parameters15: OperationParameter = {
+  parameterPath: "parameters",
+  mapper: CrossRegionRestoreJobRequestMapper
+};
+
+export const parameters16: OperationParameter = {
+  parameterPath: "parameters",
+  mapper: CrossRegionRestoreJobsRequestMapper
+};
+
+export const resourceId: OperationURLParameter = {
+  parameterPath: "resourceId",
+  mapper: {
+    serializedName: "resourceId",
+    required: true,
+    type: {
+      name: "String"
+    }
+  },
+  skipEncoding: true
+};
+
 export const jobId: OperationURLParameter = {
   parameterPath: "jobId",
   mapper: {
@@ -274,12 +316,12 @@ export const jobId: OperationURLParameter = {
   }
 };
 
-export const parameters12: OperationParameter = {
+export const parameters17: OperationParameter = {
   parameterPath: "parameters",
   mapper: AzureBackupFindRestorableTimeRangesRequestMapper
 };
 
-export const parameters13: OperationParameter = {
+export const parameters18: OperationParameter = {
   parameterPath: "parameters",
   mapper: ResourceGuardResourceMapper
 };
@@ -295,7 +337,7 @@ export const resourceGuardsName: OperationURLParameter = {
   }
 };
 
-export const parameters14: OperationParameter = {
+export const parameters19: OperationParameter = {
   parameterPath: "parameters",
   mapper: PatchResourceGuardInputMapper
 };
@@ -325,12 +367,12 @@ export const resourceGuardProxyName: OperationURLParameter = {
   }
 };
 
-export const parameters15: OperationParameter = {
+export const parameters20: OperationParameter = {
   parameterPath: "parameters",
   mapper: ResourceGuardProxyBaseResourceMapper
 };
 
-export const parameters16: OperationParameter = {
+export const parameters21: OperationParameter = {
   parameterPath: "parameters",
   mapper: UnlockDeleteRequestMapper
 };